services/ai-career-advisor/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
alembic
psycopg2-binary
openai
pinecone-client
redis
httpx
python-jose
passlib
tenacity
pytest
starlette
prometheus-client
langchain
pandas
numpy
sentence-transformers